Rename Binding to the more conventional Symbol
Note that the Greek etymology refers to a pairing.
