+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/Applications.scala
@@ -2,9 +2,9 @@ package dc10.scala.ctx.predef
 
 import cats.implicits.*
 import cats.data.StateT
-import dc10.scala.ast.Binding.Term
-import dc10.scala.ast.Binding.Term.TypeLevel.__
-import dc10.scala.ast.Binding.Term.{TypeLevel, ValueLevel}
+import dc10.scala.ast.Symbol.Term
+import dc10.scala.ast.Symbol.Term.TypeLevel.__
+import dc10.scala.ast.Symbol.Term.{TypeLevel, ValueLevel}
 import dc10.scala.ast.Statement
 import dc10.scala.ast.Statement.Expr
 import dc10.scala.ctx.ErrorF
