+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/Variables.scala
@@ -1,9 +1,9 @@
 package dc10.scala.ctx.predef
 
 import cats.data.StateT
-import dc10.scala.ast.Binding.Term
-import dc10.scala.ast.Binding.Term.{TypeLevel, ValueLevel}
-import dc10.scala.ast.Binding.Term.ValueLevel.Var.UserDefinedValue
+import dc10.scala.ast.Symbol.Term
+import dc10.scala.ast.Symbol.Term.{TypeLevel, ValueLevel}
+import dc10.scala.ast.Symbol.Term.ValueLevel.Var.UserDefinedValue
 import dc10.scala.ast.Statement
 import dc10.scala.ast.Statement.{ValDef, Expr}
 import dc10.scala.ctx.ErrorF
