+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/datatype/PrimitiveTypes.scala
@@ -1,8 +1,8 @@
 package dc10.scala.ctx.predef.datatype
 
 import cats.data.StateT
-import dc10.scala.ast.Binding.Term
-import dc10.scala.ast.Binding.Term.ValueLevel
+import dc10.scala.ast.Symbol.Term
+import dc10.scala.ast.Symbol.Term.ValueLevel
 import dc10.scala.ast.Statement
 import dc10.scala.ast.Statement.Expr
 import dc10.scala.ctx.ErrorF
