+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/file/Files.scala
@@ -1,7 +1,7 @@
 package dc10.scala.ctx.predef.file
 
 import cats.data.StateT
-import dc10.scala.ast.Binding.Package
+import dc10.scala.ast.Symbol.Package
 import dc10.scala.ast.Statement
 import dc10.scala.ctx.{ErrorF, ext}
 import dc10.scala.file.ScalaFile
